feat: add Hero component to display latest posts and events on main and about pages
